made tiles for different pipe sizes
