+++ src/main/java/gtclassic/api/pipe/GTPipeTypes.java
@@ -57,7 +57,7 @@ public class GTPipeTypes {
 		S800(800),
 		S1600(1600),
 		S3200(3200),
-		S64000(6400),
+		S6400(6400),
 		S12800(12800);
 
 		int size;
